INTEGRATION: Integration of 6305,6308 from parent to child (branch spec: dev_to_lu_dev).
[git-p4: depot-paths = "//open/lu/dev/": change = 6315]

+++ farrago/src/com/lucidera/lcs/LcsIndexSearchRel.java
@@ -225,8 +225,8 @@ class LcsIndexSearchRel extends FennelSingleRel
                 Boolean.valueOf(indexScanRel.isOrderPreserving),
                 Boolean.valueOf(isOuter), inputKeyProjObj, inputJoinProjObj,
                 inputDirectiveProjObj, 
-                (startRidParamId == null) ? 0 : startRidParamId,
-                (rowLimitParamId == null) ? 0 : rowLimitParamId
+                (startRidParamId == null) ? (Integer)0 : startRidParamId,
+                (rowLimitParamId == null) ? (Integer)0 : rowLimitParamId
             });
     }
